pyproject.toml
[project]
name = "semantic-doc-chatbot"
version = "0.1.0"
description = "A Streamlit and FastAPI chatbot that answers questions from uploaded documents using LLMs"
authors = [{ name = "Adam Fuzum Tewelde", email = "fuzome@gmail.com" }]
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "aiofiles==24.1.0",
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "python-dotenv==1.1.0",
    "pymongo==4.12.0",
    "langchain==0.3.23",
    "langchain-community==0.3.21",
    "langchain-core==0.3.51",
    "langchain-openai==0.3.12",
    "langchain-huggingface==0.1.2",
    "langchain-text-splitters==0.3.8",
    "huggingface-hub==0.30.2",
    "sentence-transformers==4.0.2",
    "streamlit==1.44.1",
    "docx2txt==0.9",
    "PyPDF2==3.0.1",
    "faiss-cpu==1.10.0",
    "aiohttp==3.11.16",
    "s3transfer==0.11.4",
    "boto3==1.37.32",
    "pydantic==2.11.3",
    "python-multipart==0.0.20",
    "typing-extensions>=4.0.0",
    "openai==1.72.0",
    "mongomock==4.3.0",
    "sentinels==1.0.0",
    "jinja2",
    "requests",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "pre-commit"
]

[project.optional-dependencies]
dev = [
    "types-requests",
    "ruff",
    "mypy",
    "types-requests",
    "types-aiofiles",
    "types-s3transfer",
    "types-awscrt"
]


[tool.ruff]
line-length = 100
exclude = ["data", ".venv"]
select = ["E", "F", "B", "I"]
ignore = ["B008"]
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "-v --tb=short --cov=app"

[tool.mypy]
python_version = "3.11"
strict = true
